+++ src/core/configmanager.h
@@ -28,10 +28,16 @@ class ConfigManager : public QObject
                          QObject *parent = nullptr);
   void save();
   void readGeneralSettings();
+
+  static bool readJsonFile(QIODevice &device, QSettings::SettingsMap &map);
+  static bool writeJsonFile(QIODevice &device,
+                            const QSettings::SettingsMap &map);
+
+  //
   QString getEditorFontFamily() const;
   void setEditorFontFamily(const QString &fontname);
-  QString getStyle() const;
-  void setStyle(const QString &stylename);
+  QString getStyleTheme() const;
+  void setStyleTheme(const QString &stylename);
   int getEditorFontSize() const;
   void setEditorFontSize(const int &fontsize);
   QString getColorTheme() const;
@@ -41,7 +47,7 @@ class ConfigManager : public QObject
   QSettings *settings;
   QString configFile;
   QString editorFontFamily;
-  QString style;
+  QString styleTheme;
   int editorFontSize;
   QString colorTheme;
 };